FROMGIT: drm/msm: Add SET_PARAM ioctl
It was always expected to have a use for this some day, so we left a
placeholder.  Now we do.  (And I expect another use in the not too
distant future when we start allowing userspace to allocate GPU iova.)

Signed-off-by: Rob Clark <robdclark@chromium.org>
Link: https://lore.kernel.org/r/20220304005317.776110-3-robdclark@gmail.com
(cherry picked from commit f7ddbf5581b474fe4a0a29244acaa1bf72234675
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:222511377
TEST=emerge-trogdor chromeos-kernel-5_15

Signed-off-by: Matt Turner <msturner@google.com>
Change-Id: I91603017a7f4bd5cf5c52b7386c66b0a48caa1e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3541291
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>