UPSTREAM: mac80211: Add support to trigger sta disconnect on hardware restart
Currently in case of target hardware restart, we just reconfig and
re-enable the security keys and enable the network queues to start
data traffic back from where it was interrupted.

Many ath10k wifi chipsets have sequence numbers for the data
packets assigned by firmware and the mac sequence number will
restart from zero after target hardware restart leading to mismatch
in the sequence number expected by the remote peer vs the sequence
number of the frame sent by the target firmware.

This mismatch in sequence number will cause out-of-order packets
on the remote peer and all the frames sent by the device are dropped
until we reach the sequence number which was sent before we restarted
the target hardware

In order to fix this, we trigger a sta disconnect, in case of target
hw restart. After this there will be a fresh connection and thereby
avoiding the dropping of frames by remote peer.

The right fix would be to pull the entire data path into the host
which is not feasible or would need lots of complex changes and
will still be inefficient.

Tested on ath10k using WCN3990, QCA6174

Signed-off-by: Youghandhar Chintala <youghand@codeaurora.org>
Link: https://lore.kernel.org/r/20220308115325.5246-2-youghand@codeaurora.org
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 7d352ccf1e9935b5222ca84e8baeb07a0c8f94b9)

BUG=b:178134516
TEST=wifi_matfunc

Cq-Depend: chromium:3774687
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I7a1897d9b6eb6dd8865f8a3bc758d16aa554919d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3774685
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>