UPSTREAM: Kbuild: add -Wno-shift-negative-value where -Wextra is used
As a preparation for moving to -std=gnu11, turn off the
-Wshift-negative-value option. This warning is enabled by gcc when
building with -Wextra for c99 or higher, but not for c89. Since
the kernel already relies on well-defined overflow behavior,
the warning is not helpful and can simply be disabled in
all locations that use -Wextra.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Jani Nikula <jani.nikula@intel.com>
Reviewed-by: Nathan Chancellor <nathan@kernel.org>
Tested-by: Sedat Dilek <sedat.dilek@gmail.com> # LLVM/Clang v13.0.0 (x86-64)
Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
(cherry picked from commit 1344794a59db2bd44b4919d2d75300fd3b1c2cd7)

BUG=b:232779979
TEST=None

Signed-off-by: Aashish Sharma <shraash@google.com>
Change-Id: Ie3b9a8e0ffd2c5b649a982637f0492129157cff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3648868
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>