UPSTREAM: ASoC: SOF: ipc-msg-injector: Use devm_kzalloc() for the rx_buffer
The rx_buffer is cleared before sending an IPC to make sure that when the
/sys/kernel/debug/sof/ipc_msg_inject file is read we will have correct
information in the buffer (no random or stale data).

But if the user reads the file before sending any message the buffer might
contain garbage which should not be interpreted.

To prevent this, clear the rx_buffer on allocation.

Fixes: cac0b0887e53 ("ASoC: SOF: Convert the generic IPC message injector into SOF client")
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Link: https://lore.kernel.org/r/20220309110104.18370-1-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 81acac8c2c88764c773811d1ec858852f81eb88a)

BUG=b:218771967
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: I75fb1390d9c6537ea0e0525efdb0028f91225bda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3623121
Tested-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>