UPSTREAM: drm/amdgpu: move VM PDEs to idle after update
Move the page tables to the idle list after updating the PDEs.

We have gone back and forth with that a couple of times because of problems
with the inter PD dependencies, but it should work now that we have the
state handling cleanly separated.

Signed-off-by: Christian König <christian.koenig@amd.com>
Acked-by: Felix Kuehling <Felix.Kuehling@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 6e97c2f9689227b3fb6e2563127669f2146b5df5)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:242065767
TEST=build test

Backported using forklift.py

Change-Id: Iadd6837aee2c7649bfabb6f8713d672a902bf475