FROMGIT: driver core: Add sysfs support for physical location of a device
When ACPI table includes _PLD fields for a device, create a new
directory (physical_location) in sysfs to share _PLD fields.

Currently without PLD information, when there are multiple of same
devices, it is hard to distinguish which device corresponds to which
physical device at which location. For example, when there are two Type
C connectors, it is hard to find out which connector corresponds to the
Type C port on the left panel versus the Type C port on the right panel.
With PLD information provided, we can determine which specific device at
which location is doing what.

_PLD output includes much more fields, but only generic fields are added
and exposed to sysfs, so that non-ACPI devices can also support it in
the future. The minimal generic fields needed for locating a device are
the following.
- panel
- vertical_position
- horizontal_position
- dock
- lid

Signed-off-by: Won Chung <wonchung@google.com>
Link: https://lore.kernel.org/r/20220314195458.271430-1-wonchung@google.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 6423d2951087231706246f81851067f7f0593d4a
https://git.kernel.org/pub/scm/linux/kernel/git/gregkh/driver-core.git
driver-core-next)

BUG=b:218506642
TEST=emerge-brya sys-kernel/chromeos-kernel-5_10

Change-Id: I534e6b1c45ef2dcd7dbcd8091b9ad9bf877d29a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3615709
Tested-by: Won Chung <wonchung@google.com>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Won Chung <wonchung@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>