UPSTREAM: thunderbolt: Add functions to support DisplayPort bandwidth allocation mode
USB4 spec defines an additional feature that DP IN adapters can
implement (alongside with the graphics DPCD register set) to support
more dynamic bandwidth management for DisplayPort tunnels. For the
connection manager the communication happens through the DP IN adapter
using a set of registers in the adapter config space allocated for this.
Add functions that export this functionality for the rest of the driver.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit e327380133d96a7a71baca65a809bf65609a1a69)

BUG=b:294519146
TEST=TBT3/4 functional tests to ensure backward compatibility

Change-Id: I42edfaa19e7dccaa357bed0885d294d81b1b1651
Signed-off-by: Jignesh Patel <jignesh1.patel@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4920349
Reviewed-by: Jignesh Patel <jignesh1.patel@intel.corp-partner.google.com>
Reviewed-by: Benson Leung <bleung@google.com>
Tested-by: Jignesh Patel <jignesh1.patel@intel.corp-partner.google.com>
Reviewed-by: Sean Paul <sean@poorly.run>