UPSTREAM: thunderbolt: Add support for DisplayPort bandwidth allocation mode
The USB4 spec defines an optional feature that allows the connection
manager to negotiate with the graphics through DPCD registers changes in
the bandwidth allocation dynamically. This is referred as "bandwidth
allocation mode" in the spec. The connection manager uses DP IN adapters
registers to communicate with the graphics, and also gets notifications
from these adapters when the graphics wants to change the bandwidth
allocation. Both the connection manager and the graphics driver needs to
support this.

We check if the DP IN adapter supports this and if it does enable it
before establishing a DP tunnel. Then we react on DP_BW notifications
coming from the DP IN adapter and update the bandwidth allocation
accordingly (within the maximum common capabilities the DP IN/OUT
support).

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 6ce3563520be90a155706bafc186fc264a13850e)

BUG=b:294519146
TEST=TBT3/4 functional tests to ensure backward compatibility

Change-Id: I62f23105544eb148b419805390303a3d06e487e0
Signed-off-by: Jignesh Patel <jignesh1.patel@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4920343
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Jignesh Patel <jignesh1.patel@intel.corp-partner.google.com>
Tested-by: Jignesh Patel <jignesh1.patel@intel.corp-partner.google.com>