UPSTREAM: thunderbolt: Replace usage of found with dedicated list iterator variable
To move the list iterator variable into the list_for_each_entry_*()
macro in the future it should be avoided to use the list iterator
variable after the loop body.

To *never* use the list iterator variable after the loop it was
concluded to use a separate iterator variable instead of a
found boolean [1].

This removes the need to use a found variable and simply checking if
the variable was set, can determine if the break/goto was hit.

Link: https://lore.kernel.org/all/CAHk-=wgRr_D8CB-D9Kg-c=EHreAsk5SqXPwr9Y7k9sA6cWXJ6w@mail.gmail.com/
Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 03941ed91c7231e4973fb50de6c349974405df4e
 https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:242112030
TEST=emerge-rex sys-kernel/chromeos-kernel-5_15

Signed-off-by: Narasimha Rao Rayala <narasimha.rao.rayala@intel.com>
Change-Id: I26d0b038bbc9f5640f4a605710a21af31a94da6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3887004
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>