FROMGIT: ASoC: SOF: ipc: Use msg->reply_data directly in snd_sof_ipc_get_reply()
Instead of using a local reply to first read out the header from the
mailbox then memcpy it or read it again to msg->reply_data, read it
directly to it's final place from the start.

If we received an error we do not need to do a memcpy anymore.
If the reply is reporting a success then we don not need to read the reply
again from the mailbox if the reply_size equals to the already read header
size.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Link: https://lore.kernel.org/r/20220330201926.1330402-9-ranjani.sridharan@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 5db8eb5b9e35c712cc63fcebf04c80ace9812961
 https://kernel.googlesource.com/pub/scm/linux/kernel/git/broonie/sound.git for-next)

BUG=b:218771967
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: Ie5ac7ee981545c3b8c5fc6b7d96bb21e3e59d195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3629454
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>