UPSTREAM: ASoC: soc-utils: Add kunit test for snd_soc_tdm_params_to_bclk()
Create a new kunit test for soc-utils and use it to test
snd_soc_tdm_params_to_bclk().

The test uses a table of values to avoid the possibility that an
on-the-fly generator contains the same algorithmic error as the
function-under-test and so fails to detect a bug.

There is no need to test every possible combination of values.
Enough test cases are included to give confidence that the function
is producing the correct results.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20220405135419.1230088-4-rf@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 89342fa38bbaade51584f255eee5cd43621f4e10)

BUG=b:218700194
TEST=Test Audio use cases

Signed-off-by: SHANKAR BOYANE <shankarx.boyane@intel.com>
Change-Id: I5642aa622a66ea0051450b6527b6b2f29b4616b8