UPSTREAM: device property: Add helper to match multiple connections
In some cases multiple connections with the same connection id
needs to be resolved from a fwnode graph.

One such example is when separate hardware is used for performing muxing
and/or orientation switching of the SuperSpeed and SBU lines in a USB
Type-C connector. In this case the connector needs to belong to a graph
with multiple matching remote endpoints, and the Type-C controller needs
to be able to resolve them both.

Add a new API that allows this kind of lookup.

Reviewed-by: Andy Shevchenko <andy.shevchenko@gmail.com>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20220422222351.1297276-2-bjorn.andersson@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 7a20917d30fb627bebff6bacbe860ad4eb3a04d6)

BUG=b:233677023
TEST=Tested alternate mode entry on ARM device using anx7625

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I2584483c91f6a2e8b4341f35b65a1fe10973f2c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3749696
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>