UPSTREAM: usb: typec: mux: Allow multiple mux_devs per mux
In the Qualcomm platforms the USB/DP PHY handles muxing and orientation
switching of the SuperSpeed lines, but the SBU lines needs to be
connected and switched by external (to the SoC) hardware.

It's therefor necessary to be able to have the TypeC controller operate
multiple TypeC muxes and switches. Use the newly introduced indirection
object to handle this, to avoid having to taint the TypeC controllers
with knowledge about the downstream hardware configuration.

The max number of devs per indirection is set to 3, which account for
being able to mux/switch the USB HS, SS and SBU lines, as per defined
defined in the usb-c-connector binding. This number could be grown if
need arrises at a later point in time.

Acked-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20220422222351.1297276-6-bjorn.andersson@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 71793b579ba6825d221445e3cfcd50da427b4f0e)

BUG=b:233677023
TEST=Tested alternate mode entry on ARM device using anx7625

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I813655ceb08c4596857f973db55cf267fe7cd834
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3749699
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>