CHROMIUM: iwl7000: mvm: refactor __iwl_mvm_unassign_vif_chanctx()
Since parts of the functionality of this function is going to be used
also by the MLD version of it, put in a separate function the parts
that are common for both MLD and non-MLD modes.
The common function will later be used in the MLD ops.

Change-Id: I2222dc86cf7d7a7bb58c6a2f2529c8089bfe58b4
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 775deec80b5e7b69496441e51add55bd9a897d38