FROMGIT: ASoC: SOF: ipc3-topology: Correct get_control_data for non bytes payload
It is possible to craft a topology where sof_get_control_data() would do
out of bounds access because it expects that it is only called when the
payload is bytes type.
Confusingly it also handles other types of controls, but the payload
parsing implementation is only valid for bytes.

Fix the code to count the non bytes controls and instead of storing a
pointer to sof_abi_hdr in sof_widget_data (which is only valid for bytes),
store the pointer to the data itself and add a new member to save the size
of the data.

In case of non bytes controls we store the pointer to the chanv itself,
which is just an array of values at the end.

In case of bytes control, drop the wrong cdata->data (wdata[i].pdata) check
against NULL since it is incorrect and invalid in this context.
The data is pointing to the end of cdata struct, so it should never be
null.

Reported-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Link: https://lore.kernel.org/r/20220427185221.28928-1-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit a962890a5a3cce903ff7c7a19fadee63ed9efdc7
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-5.19)

BUG=b:230814364
TEST=none

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Change-Id: I09c7d13bdbc0f1c2b9c1ccf539c885efedd85927
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3611383
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>